src/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.1
django-constance==3.1.0
django-picklefield==3.2
django-widget-tweaks==1.5.0
dynamic-django-forms==0.1.11
fonttools==4.53.1
fpdf2==2.7.9
google-api-core==2.16.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gspread==6.0.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
lxml==5.2.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pillow==10.3.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pypandoc==1.13
pyparsing==3.1.1
pypng==0.20220715.0
python-docx==1.1.2
python3-openid==3.2.0
qrcode==7.4.2
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
typing_extensions==4.10.0
urllib3==2.2.0
whitenoise==6.6.0
